% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomDeMorarApp.R
\name{app_server}
\alias{app_server}
\title{Bom de Morar Server}
\usage{
app_server()
}
\description{
Backend for the Proof of Concept Bom de Morar App.
}
\details{
This function only provides the backend of the application.
}
