% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomDeMorarApp.R
\name{app_ui}
\alias{app_ui}
\title{Bom de Morar UI}
\usage{
app_ui()
}
\value{
the HTML for the frontend.
}
\description{
User Interface (UI) for the Proof of Concept Bom de Morar App.
}
\details{
This function only provides  the User Interface (frontend)
of the application.
}
