% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkAnyAdultFamily}
\alias{checkAnyAdultFamily}
\title{Checar adulto na família}
\usage{
checkAnyAdultFamily(family)
}
\arguments{
\item{family}{base de dados com informações de todos os membros da família.}
}
\description{
Essa função irá checar se há algum membro na família com idade acima de 
18 anos a partir das informações do CadÚnico.
}
