% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkAuxilioAcolhida_CPF}
\alias{checkAuxilioAcolhida_CPF}
\title{Checar Recebimento de Auxílio Acolhida}
\usage{
checkAuxilioAcolhida_CPF(family, file_path = "data/auxilioAcolhida/")
}
\arguments{
\item{family}{base de dados com informações de todos os membros da família.}

\item{file_path}{caminho para o arquivo da base de auxilio moradia.}
}
\description{
Essa função irá checar se há algum membro da família que recebe benefício do 
Programa Auxílio Acolhida (Antigo Aluguel Social):
}
\details{
Essa função realizará uma operação de `join` junto a base do programa Auxílio
Acolhida utilizando a informação do número `CPF` do indivíduo.
}
