% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkAuxilioMoradia_NIS}
\alias{checkAuxilioMoradia_NIS}
\title{Checar recebimento de auxílio Moradia}
\usage{
checkAuxilioMoradia_NIS(family, file_path = "data/auxilioMoradia/")
}
\arguments{
\item{family}{base de dados com informações de todos os membros da família.}

\item{file_path}{caminho para o arquivo da base de auxilio moradia.}
}
\description{
Essa função irá checar se há algum membro da família que recebe benefício do 
Programa Auxílio Moradia.
}
\details{
Essa função realizará uma operação de `join` junto a base do programa Auxílio
Moradia utilizando a informação do CPF do indivíduo.
}
