% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkCAD_CPF}
\alias{checkCAD_CPF}
\title{Função para checar inscrição no CadUnico}
\usage{
checkCAD_CPF(cpf, file_path = "data/cadunico/")
}
\arguments{
\item{cpf}{caracter com o número do CPF.}

\item{file_path}{endereço do arquivo da base do CadÚnico. Default para "data/cadunico/.}
}
\description{
Essa função irá analisar se o CPF informado está inscrito no CadUnico.
}
