% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkCAD_NIS}
\alias{checkCAD_NIS}
\title{Função para checar inscrição no CadUnico}
\usage{
checkCAD_NIS(nis, file_path = "data/cadunico/")
}
\arguments{
\item{nis}{caracter com o número do NIS.}

\item{file_path}{endereço do arquivo da base do CadÚnico. Default para "data/cadunico/.}
}
\description{
Essa função irá analisar se o NIS informado está inscrito no CadUnico.
}
