% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkCadastroAtualizadoCAD}
\alias{checkCadastroAtualizadoCAD}
\title{Checar Atualização de Cadastro no CadÚnico}
\usage{
checkCadastroAtualizadoCAD(
  dtaAtualizacaoCadastral,
  dtaRefCad,
  updated_within_months = 24
)
}
\arguments{
\item{dtaAtualizacaoCadastral}{data da última atualização cadastral}

\item{dtaRefCad}{data de referência do CAD}

\item{updated_within_months}{número de meses para checar atualização.
Default para 24 meses.}
}
\description{
Função para checar se o cadastro foi atualizado nos últimos 24 meses:
}
