% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkElegibility_CPF}
\alias{checkElegibility_CPF}
\title{Checar Elegibilidade do indivíduo.}
\usage{
checkElegibility_CPF(cpf)
}
\arguments{
\item{cpf}{número de cpf do indivíduos}
}
\description{
Função para checar elegibilidade ao programa utilizando o CPF:
}
