% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkElegibility_NIS}
\alias{checkElegibility_NIS}
\title{Checar Elegibilidade do indivíduo.}
\usage{
checkElegibility_NIS(nis)
}
\arguments{
\item{nis}{número de `NIS` do indivíduo}
}
\description{
Função para checar elegibilidade ao programa utilizando o `NIS`.
}
