% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getFamilyCAD}
\alias{getFamilyCAD}
\title{Função para buscar informações de famílias no CadÚnico}
\usage{
getFamilyCAD(cod_familiar, file_path = "data/cadunico")
}
\arguments{
\item{cod_familiar}{caractere com o código de identificação da família no CadÚnico}

\item{file_path}{endereço do arquivo da base do CadÚnico. Default para "data/cadunico/.}
}
\description{
Essa função irá buscar as informações de todos membros da família na base 
do CadÚnico a partir de seu código de identificação.
}
