% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getInfoCAD_CPF}
\alias{getInfoCAD_CPF}
\title{Função para buscar informação de indivíduos no CadÚnico}
\usage{
getInfoCAD_CPF(cpf, file_path = "data/cadunico/")
}
\arguments{
\item{cpf}{caractere com número do CPF do indivíduo. Esse número possui 11 digitos.}

\item{file_path}{endereço do arquivo da base do CadÚnico. Default para "data/cadunico/.}
}
\description{
Essa função irá buscar as informações socio demográficas do CPF na base do CadÚnico.
}
