% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getInfoCAD_NIS}
\alias{getInfoCAD_NIS}
\title{Função para buscar informação de indivíduos no CadÚnico}
\usage{
getInfoCAD_NIS(nis, file_path = "data/cadunico/")
}
\arguments{
\item{nis}{caractere com número do CPF do indivíduo. Esse número possui 11 digitos.}

\item{file_path}{endereço do arquivo da base do CadÚnico. Default para "data/cadunico/.}
}
\description{
Essa função irá buscar as informações socio demográficas do NIS na base do CadÚnico.
}
