% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{listaImoveisDisp}
\alias{listaImoveisDisp}
\title{Lista de Imóveis Disponíveis}
\usage{
listaImoveisDisp(cr, aluguel_up, aluguel_lw, file_path = "data/imoveis/")
}
\description{
Função para selecionar os imóveis possíveis, dado o nível de comprometimento
 de renda da família.
 
 @param cr comprometimento de renda
 @param aluguel_up valor máximo de aluguel
 @param aluguel_lw valor mínumo de aluguel
 @param file_path path  do banco de dados de imóveis
 
 @export
}
