% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{playProgramaBomDeMorar}
\alias{playProgramaBomDeMorar}
\title{Função para Calcular Subsídio.}
\usage{
playProgramaBomDeMorar(infoFamily, SM, SUBSIDIO_MIN, SUBSIDIO_MAX, ALUGUEL_MAX)
}
\arguments{
\item{infoFamily}{dataframe com as informações da família extraídas do CadUnico}

\item{SM}{Valor do salário mínimo}

\item{SUBSIDIO_MIN}{valor do subsídio mínimo}

\item{SUBSIDIO_MAX}{valor do subsídio máximo}

\item{ALGUEL_MAX}{valor do aluguel máximo}
}
\description{
Essa função para calcular a modalidade do programa e valor do subsídio.
}
