% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plotMap}
\alias{plotMap}
\title{Plotar mapa com imóveis}
\usage{
plotMap(imoveis)
}
\arguments{
\item{imoveis}{banco de dados com os imóveis}
}
\description{
Essa função irá plotr o mapa com os imóveis.
}
